make the simulation step slider re-usable
